pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cv_to_role_analyzer"
version = "0.1.0"
description = "Analyzes CVs against job descriptions."
authors = [{name = "Your Name", email = "your.email@example.com"}]
dependencies = [
    "click",
    "google-generativeai",
    "python-dotenv",
    "pytest",
    "pypdf"
]

[tool.pytest.ini_options]
testpaths = ["tests"]